fix sorting in all projects table
